swtiched to ver2 as default version
